implement stream dis

+++ rwkv_ops/mhc_kernel/common_kernel/kernels/stream_aggregate.cuh
@@ -22,6 +22,7 @@ __global__ void stream_aggregate_fwd_kernel(
     const float* __restrict__ H_pre,
     int64_t BT, int n, int64_t C) {
     
+    // blockIdx.x 强转为 int64_t，防止 32 位溢出
     int64_t btc = (int64_t)blockIdx.x * blockDim.x + threadIdx.x;
     if (btc >= BT * C) return;
 
@@ -34,9 +35,11 @@ __global__ void stream_aggregate_fwd_kernel(
     #pragma unroll
     for (int i = 0; i < 8; i++) { // 假设 n <= 8，由 mHC 论文设定
         if (i < n) {
-            float h_val = PER_TOKEN_H ? H_pre[bt * n + i] : H_pre[i];
+            // [修改]: bt * n 可能超过 32 位，且 n 为 int，显式强转 n 为 int64_t
+            float h_val = PER_TOKEN_H ? H_pre[bt * (int64_t)n + i] : H_pre[i];
             // 关键：读取 bf16 后立即转为 fp32 参与运算
-            float x_val = to_float(inp[bt * n * C + (int64_t)i * C + c]);
+            // [修改]: 显式强转 n 为 int64_t，确保 bt * n * C 全程为 64 位运算
+            float x_val = to_float(inp[bt * (int64_t)n * C + (int64_t)i * C + c]);
             sum += h_val * x_val;
         }
     }
@@ -54,15 +57,18 @@ __global__ void stream_aggregate_bwd_dx_kernel(
     const float* __restrict__ H_pre,
     int64_t BT, int n, int64_t C) {
     
+    // blockIdx.x 强转为 int64_t
     int64_t btc = (int64_t)blockIdx.x * blockDim.x + threadIdx.x;
     int i = blockIdx.y; 
 
     if (btc < BT * C && i < n) {
         int64_t bt = btc / C;
-        float h_val = PER_TOKEN_H ? H_pre[bt * n + i] : H_pre[i];
+        // [修改]: bt * n 增加 (int64_t)n 强转
+        float h_val = PER_TOKEN_H ? H_pre[bt * (int64_t)n + i] : H_pre[i];
         float grad_val = d_out[btc]; // 接收 FP32 梯度
         
-        d_inp[bt * n * C + (int64_t)i * C + (btc % C)] = to_bf(grad_val * h_val);
+        // [修改]: bt * n * C 增加 (int64_t)n 强转
+        d_inp[bt * (int64_t)n * C + (int64_t)i * C + (btc % C)] = to_bf(grad_val * h_val);
     }
 }
 
@@ -80,7 +86,8 @@ __global__ void stream_aggregate_bwd_dh_kernel(
     cg::thread_block block = cg::this_thread_block();
     cg::thread_block_tile<32> warp = cg::tiled_partition<32>(block);
 
-    int64_t bt = blockIdx.x;
+    // blockIdx.x 在 C++ 侧对应 BT，若 BT 很大这里 int64_t 是必要的
+    int64_t bt = blockIdx.x; 
     int i = blockIdx.y;
     if (bt >= BT || i >= n) return;
 
@@ -89,7 +96,8 @@ __global__ void stream_aggregate_bwd_dh_kernel(
     float thread_sum = 0.0f;
     for (int64_t c = threadIdx.x; c < C; c += BLOCK_SIZE) {
         float g_val = d_out[bt * C + c]; 
-        float x_val = to_float(inp[bt * n * C + (int64_t)i * C + c]);
+        // [修改]: bt * n * C 增加 (int64_t)n 强转
+        float x_val = to_float(inp[bt * (int64_t)n * C + (int64_t)i * C + c]);
         thread_sum += g_val * x_val;
     }
 
@@ -103,12 +111,15 @@ __global__ void stream_aggregate_bwd_dh_kernel(
     if (warp_id == 0) {
         float val = (lane_id < (BLOCK_SIZE / 32)) ? s_reduce[lane_id] : 0.0f;
         float block_sum = cg::reduce(warp, val, cg::plus<float>());
-        if (lane_id == 0) d_H_pre[bt * n + i] = block_sum;
+        // [修改]: bt * n 增加 (int64_t)n 强转
+        if (lane_id == 0) d_H_pre[bt * (int64_t)n + i] = block_sum;
     }
 }
 
 inline void stream_aggregate_forward(floatX* out, const floatX* inp, const float* H_pre, int64_t BT, int n, int64_t C, bool per_token, cudaStream_t stream) {
     dim3 threads(256);
+    // 注意: BT * C 如果极大，超过 uint32 范围，dim3.x 会截断，这是 CUDA 硬件限制。
+    // 但计算过程本身 BT*C 是 int64 不会溢出。
     dim3 blocks((BT * C + 255) / 256);
     if (per_token) stream_aggregate_fwd_kernel<true><<<blocks, threads, 0, stream>>>(out, inp, H_pre, BT, n, C);
     else stream_aggregate_fwd_kernel<false><<<blocks, threads, 0, stream>>>(out, inp, H_pre, BT, n, C);
